refactor: extract config_server.py into focused web interface modules
