+++ src/dicton/interfaces/config_server.py
@@ -1,465 +1,76 @@
-"""Configuration web UI server for Dicton."""
+"""Configuration web UI server for Dicton.
 
-import base64
-import json
-import os
-import shutil
+This module is the public entry-point.  The heavy lifting now lives in
+dedicated submodules under ``dicton.interfaces.web``:
+
+* ``env_io``       – .env file reading / writing
+* ``config_logic`` – field maps, status checks, dictionary helpers
+* ``templates``    – HTML / asset loading
+
+Everything is re-exported here so that existing callers (CLI, tests) that do
+``from dicton.interfaces.config_server import X`` keep working.
+"""
+
+from __future__ import annotations
+
 import webbrowser
-from pathlib import Path
 from threading import Timer
-from typing import Any
 
-from ..shared.config import Config, config
-from ..shared.platform_utils import IS_LINUX, IS_MACOS, IS_WINDOWS, get_platform_info
+from ..shared.config import Config as Config  # noqa: F401
+
+# -- re-exports used by tests for monkeypatching -----------------------------
+from ..shared.platform_utils import get_platform_info as get_platform_info  # noqa: F401
+from ..shared.startup import (  # noqa: F401
+    get_autostart_state as get_autostart_state,
+)
 from ..shared.startup import (
-    get_autostart_state,
-    has_display_session,
-    launch_background,
-    set_autostart,
+    has_display_session as has_display_session,
+)
+from ..shared.startup import (
+    launch_background as launch_background,
+)
+from ..shared.startup import (
+    set_autostart as set_autostart,
 )
 
-
-def _load_logo_base64() -> str:
-    """Load logo from package assets folder and convert to base64."""
-    logo_path = Path(__file__).parent.parent / "assets" / "logo.png"
-    if logo_path.exists():
-        return base64.b64encode(logo_path.read_bytes()).decode("utf-8")
-    return ""
-
-
-LOGO_BASE64 = _load_logo_base64()
-
-
-def _load_html_template(template_name: str) -> str:
-    """Load HTML template for the configuration UI."""
-    template_path = Path(__file__).parent.parent / "assets" / template_name
-    if template_path.exists():
-        html = template_path.read_text(encoding="utf-8")
-        return html.replace("{{LOGO_BASE64}}", LOGO_BASE64)
-    return "<html><body>Config UI template missing.</body></html>"
-
-
-SETUP_HTML_TEMPLATE = _load_html_template("setup_ui.html")
-ADVANCED_HTML_TEMPLATE = _load_html_template("config_ui.html")
-
-CONFIG_FIELD_MAP = {
-    "stt_provider": "STT_PROVIDER",
-    "mistral_api_key": "MISTRAL_API_KEY",
-    "elevenlabs_api_key": "ELEVENLABS_API_KEY",
-    "gemini_api_key": "GEMINI_API_KEY",
-    "anthropic_api_key": "ANTHROPIC_API_KEY",
-    "llm_provider": "LLM_PROVIDER",
-    "enable_advanced_modes": "ENABLE_ADVANCED_MODES",
-    "theme_color": "THEME_COLOR",
-    "visualizer_style": "VISUALIZER_STYLE",
-    "animation_position": "ANIMATION_POSITION",
-    "visualizer_backend": "VISUALIZER_BACKEND",
-    "hotkey_base": "HOTKEY_BASE",
-    "hotkey_double_tap_window_ms": "HOTKEY_DOUBLE_TAP_WINDOW_MS",
-    "filter_fillers": "FILTER_FILLERS",
-    "enable_reformulation": "ENABLE_REFORMULATION",
-    "language": "LANGUAGE",
-    "debug": "DEBUG",
-    "custom_hotkey_value": "CUSTOM_HOTKEY_VALUE",
-    "secondary_hotkey": "SECONDARY_HOTKEY",
-    "secondary_hotkey_translation": "SECONDARY_HOTKEY_TRANSLATION",
-    "secondary_hotkey_act_on_text": "SECONDARY_HOTKEY_ACT_ON_TEXT",
-    "mute_playback_on_recording": "MUTE_PLAYBACK_ON_RECORDING",
-    "playback_mute_strategy": "PLAYBACK_MUTE_STRATEGY",
-    "mute_backend": "MUTE_BACKEND",
-}
-
-CONFIG_BOOL_FIELDS = {
-    "enable_advanced_modes",
-    "filter_fillers",
-    "enable_reformulation",
-    "debug",
-    "mute_playback_on_recording",
-}
-CONFIG_STRING_FIELDS = set(CONFIG_FIELD_MAP) - CONFIG_BOOL_FIELDS
-
-_setup_state = {
-    "first_test_passed": False,
-    "last_test_text": "",
-}
-
-
-def get_env_path() -> Path:
-    """Get the .env file path for writing (always user config dir)."""
-    return Config.CONFIG_DIR / ".env"
-
-
-def _find_env_file() -> Path | None:
-    """Find .env file - prioritize user config over system config."""
-    locations = [
-        Config.CONFIG_DIR / ".env",  # User config dir (~/.config/dicton/) - FIRST!
-        Path.cwd() / ".env",  # Current working directory
-        Path("/opt/dicton/.env"),  # System install (read-only fallback)
-    ]
-    for env_path in locations:
-        if env_path.exists():
-            return env_path
-    return None
-
-
-def read_env_file() -> dict[str, str]:
-    """Read the .env file and return as dict (checks multiple locations)."""
-    env_path = _find_env_file()
-    if env_path is None:
-        return {}
-
-    env_vars: dict[str, str] = {}
-    with open(env_path, encoding="utf-8") as f:
-        for line in f:
-            line = line.strip()
-            if not line or line.startswith("#"):
-                continue
-            if "=" in line:
-                key, _, value = line.partition("=")
-                # Strip quotes from value
-                value = value.strip().strip("\"'")
-                env_vars[key.strip()] = value
-
-    return env_vars
-
-
-def write_env_file(env_vars: dict[str, str]) -> None:
-    """Write env vars to .env file."""
-    env_path = get_env_path()
-    env_path.parent.mkdir(parents=True, exist_ok=True)
-
-    lines = ["# Dicton configuration", "# Generated by dicton --config", ""]
-
-    for key, value in sorted(env_vars.items()):
-        # Quote values with spaces
-        if " " in value or not value:
-            value = f'"{value}"'
-        lines.append(f"{key}={value}")
-
-    with open(env_path, "w", encoding="utf-8") as f:
-        f.write("\n".join(lines) + "\n")
-
-
-def _mask_api_key(key: str) -> str:
-    """Mask API key showing first char + dots + last 2 chars."""
-    if not key or len(key) < 4:
-        return ""
-    return f"{key[0]}{'•' * 8}{key[-2:]}"
-
-
-def get_current_config() -> dict[str, Any]:
-    """Get current configuration as dict."""
-    env_vars = read_env_file()
-
-    # Get API keys with masking
-    mistral_key = env_vars.get("MISTRAL_API_KEY", "")
-    elevenlabs_key = env_vars.get("ELEVENLABS_API_KEY", "")
-    gemini_key = env_vars.get("GEMINI_API_KEY", "")
-    anthropic_key = env_vars.get("ANTHROPIC_API_KEY", "")
-
-    return {
-        # STT settings
-        "stt_provider": env_vars.get("STT_PROVIDER", config.STT_PROVIDER),
-        # API keys - masked values for display
-        "mistral_api_key_set": bool(mistral_key),
-        "mistral_api_key_masked": _mask_api_key(mistral_key),
-        "elevenlabs_api_key_set": bool(elevenlabs_key),
-        "elevenlabs_api_key_masked": _mask_api_key(elevenlabs_key),
-        "gemini_api_key_set": bool(gemini_key),
-        "gemini_api_key_masked": _mask_api_key(gemini_key),
-        "anthropic_api_key_set": bool(anthropic_key),
-        "anthropic_api_key_masked": _mask_api_key(anthropic_key),
-        # Other config values
-        "llm_provider": env_vars.get("LLM_PROVIDER", config.LLM_PROVIDER),
-        "enable_advanced_modes": (env_vars.get("ENABLE_ADVANCED_MODES", "false").lower() == "true"),
-        "theme_color": env_vars.get("THEME_COLOR", config.THEME_COLOR),
-        "visualizer_style": env_vars.get("VISUALIZER_STYLE", config.VISUALIZER_STYLE),
-        "animation_position": env_vars.get("ANIMATION_POSITION", config.ANIMATION_POSITION),
-        "visualizer_backend": env_vars.get("VISUALIZER_BACKEND", config.VISUALIZER_BACKEND),
-        "hotkey_base": env_vars.get("HOTKEY_BASE", config.HOTKEY_BASE),
-        "hotkey_double_tap_window_ms": env_vars.get(
-            "HOTKEY_DOUBLE_TAP_WINDOW_MS", str(config.HOTKEY_DOUBLE_TAP_WINDOW_MS)
-        ),
-        "filter_fillers": env_vars.get("FILTER_FILLERS", "true").lower() == "true",
-        "enable_reformulation": env_vars.get("ENABLE_REFORMULATION", "false").lower() == "true",
-        "language": env_vars.get("LANGUAGE", config.LANGUAGE),
-        "debug": env_vars.get("DEBUG", "false").lower() == "true",
-        # Hotkey settings
-        "custom_hotkey_value": env_vars.get("CUSTOM_HOTKEY_VALUE", "alt+g"),
-        "secondary_hotkey": env_vars.get("SECONDARY_HOTKEY", "none"),
-        "secondary_hotkey_translation": env_vars.get("SECONDARY_HOTKEY_TRANSLATION", "none"),
-        "secondary_hotkey_act_on_text": env_vars.get("SECONDARY_HOTKEY_ACT_ON_TEXT", "none"),
-        "mute_playback_on_recording": (
-            env_vars.get("MUTE_PLAYBACK_ON_RECORDING", "true").lower() == "true"
-        ),
-        "playback_mute_strategy": env_vars.get(
-            "PLAYBACK_MUTE_STRATEGY", config.PLAYBACK_MUTE_STRATEGY
-        ),
-        "mute_backend": env_vars.get("MUTE_BACKEND", config.MUTE_BACKEND),
-    }
-
-
-def save_config(data: dict[str, Any]) -> None:
-    """Save configuration to .env file."""
-    env_vars = read_env_file()
-
-    for ui_field, env_var in CONFIG_FIELD_MAP.items():
-        if ui_field in data:
-            value = data[ui_field]
-            if isinstance(value, bool):
-                value = "true" if value else "false"
-            env_vars[env_var] = str(value)
-            os.environ[env_var] = str(value)
-
-    write_env_file(env_vars)
-
-    Config.reload_config()
-
-    from ..adapters.stt.factory import clear_provider_cache
-
-    clear_provider_cache()
-
-    from ..adapters.llm.factory import cleanup as llm_cleanup
-
-    llm_cleanup()
-
-
-def _get_env_string(env_vars: dict[str, str], key: str, default: str = "") -> str:
-    return env_vars.get(key, default).strip()
-
-
-def _get_env_bool(env_vars: dict[str, str], key: str, default: bool = False) -> bool:
-    if key not in env_vars:
-        return default
-    return env_vars[key].strip().lower() == "true"
-
-
-def _stt_status(env_vars: dict[str, str]) -> dict[str, Any]:
-    selected = _get_env_string(env_vars, "STT_PROVIDER", config.STT_PROVIDER).lower()
-    has_mistral = bool(_get_env_string(env_vars, "MISTRAL_API_KEY"))
-    has_elevenlabs = bool(_get_env_string(env_vars, "ELEVENLABS_API_KEY"))
-
-    available = []
-    if has_mistral:
-        available.append("mistral")
-    if has_elevenlabs:
-        available.append("elevenlabs")
-
-    if selected == "mistral":
-        ready = has_mistral
-        detail = "Mistral API key saved." if ready else "Add a Mistral API key."
-    elif selected == "elevenlabs":
-        ready = has_elevenlabs
-        detail = "ElevenLabs API key saved." if ready else "Add an ElevenLabs API key."
-    else:
-        ready = bool(available)
-        if ready:
-            detail = f"Auto mode can use: {', '.join(name.title() for name in available)}."
-        else:
-            detail = "Add a Mistral or ElevenLabs API key."
-
-    return {
-        "ready": ready,
-        "selected": selected,
-        "available": available,
-        "detail": detail,
-    }
-
-
-def _hotkey_status(env_vars: dict[str, str]) -> dict[str, Any]:
-    hotkey_base = _get_env_string(env_vars, "HOTKEY_BASE", config.HOTKEY_BASE).lower()
-    custom_hotkey = _get_env_string(
-        env_vars, "CUSTOM_HOTKEY_VALUE", config.CUSTOM_HOTKEY_VALUE
-    ).lower()
-
-    if IS_LINUX:
-        if hotkey_base in {"fn", "custom"}:
-            try:
-                import evdev  # noqa: F401
-            except ImportError:
-                return {
-                    "ready": False,
-                    "mode": hotkey_base,
-                    "detail": "Linux FN/custom hotkeys require the evdev dependency.",
-                }
-
-            readable_devices = any(
-                os.access(path, os.R_OK) for path in Path("/dev/input").glob("event*")
-            )
-            if not readable_devices:
-                return {
-                    "ready": False,
-                    "mode": hotkey_base,
-                    "detail": "Dicton cannot read /dev/input yet. Log out and back in after installation or add your user to the input group.",
-                }
-
-            if hotkey_base == "custom" and not custom_hotkey:
-                return {
-                    "ready": False,
-                    "mode": hotkey_base,
-                    "detail": "Choose a custom hotkey before continuing.",
-                }
-
-            return {
-                "ready": True,
-                "mode": hotkey_base,
-                "detail": "Hotkey backend is ready.",
-            }
-
-        if os.environ.get("DISPLAY"):
-            return {
-                "ready": True,
-                "mode": hotkey_base,
-                "detail": "Legacy X11 hotkey backend is available.",
-            }
-
-        return {
-            "ready": False,
-            "mode": hotkey_base,
-            "detail": "Legacy hotkeys require an X11 session. Prefer FN or custom mode on Linux.",
-        }
-
-    if hotkey_base == "custom" and not custom_hotkey:
-        return {
-            "ready": False,
-            "mode": hotkey_base,
-            "detail": "Choose a custom hotkey before continuing.",
-        }
-
-    return {
-        "ready": True,
-        "mode": hotkey_base,
-        "detail": "Hotkey backend will use the packaged keyboard listener.",
-    }
-
-
-def _text_output_status() -> dict[str, Any]:
-    if IS_LINUX:
-        if shutil.which("xdotool"):
-            return {
-                "ready": True,
-                "detail": "xdotool is installed for text insertion.",
-            }
-        return {
-            "ready": False,
-            "detail": "Install xdotool for reliable text insertion on Linux.",
-        }
-
-    if IS_WINDOWS:
-        return {"ready": True, "detail": "Windows text insertion backend is available."}
-
-    if IS_MACOS:
-        return {
-            "ready": True,
-            "detail": "macOS text insertion backend uses the fallback keyboard driver.",
-        }
-
-    return {"ready": False, "detail": "Unsupported platform."}
-
-
-def _llm_status(env_vars: dict[str, str]) -> dict[str, Any]:
-    """Check whether an LLM provider key is configured for translation."""
-    provider = env_vars.get("LLM_PROVIDER", config.LLM_PROVIDER).lower()
-    gemini_key = env_vars.get("GEMINI_API_KEY", "")
-    anthropic_key = env_vars.get("ANTHROPIC_API_KEY", "")
-
-    if provider == "gemini" and gemini_key:
-        return {"ready": True, "detail": "Gemini API key saved for translation."}
-    if provider == "anthropic" and anthropic_key:
-        return {"ready": True, "detail": "Anthropic API key saved for translation."}
-    if gemini_key or anthropic_key:
-        return {"ready": True, "detail": "LLM API key saved (translation available)."}
-
-    return {
-        "ready": False,
-        "detail": "Set a Gemini or Anthropic API key to enable translation.",
-    }
-
-
-def build_setup_status() -> dict[str, Any]:
-    """Build the setup/readiness status for the onboarding UI."""
-    env_vars = read_env_file()
-    stt = _stt_status(env_vars)
-    hotkey = _hotkey_status(env_vars)
-    output = _text_output_status()
-    llm = _llm_status(env_vars)
-    autostart = get_autostart_state()
-    launch_ready = has_display_session()
-
-    if not stt["ready"]:
-        next_step = "speech"
-    elif not hotkey["ready"]:
-        next_step = "hotkey"
-    elif autostart["supported"] and not autostart["enabled"]:
-        next_step = "autostart"
-    elif not _setup_state["first_test_passed"]:
-        next_step = "test"
-    else:
-        next_step = "done"
-
-    return {
-        "platform": get_platform_info(),
-        "config_path": str(get_env_path()),
-        "config_saved": get_env_path().exists(),
-        "first_test_passed": _setup_state["first_test_passed"],
-        "last_test_text": _setup_state["last_test_text"],
-        "launch_ready": launch_ready,
-        "launch_detail": (
-            "Desktop session detected." if launch_ready else "Start Dicton from a desktop session."
-        ),
-        "checks": {
-            "stt": stt,
-            "llm": llm,
-            "hotkey": hotkey,
-            "text_output": output,
-            "autostart": autostart,
-        },
-        "next_step": next_step,
-    }
-
-
-def get_dictionary() -> dict[str, Any]:
-    """Get dictionary contents."""
-    dictionary_path = Config.CONFIG_DIR / "dictionary.json"
-    if not dictionary_path.exists():
-        return {"similarity_words": [], "replacements": {}, "case_sensitive": {}, "patterns": []}
-
-    try:
-        with open(dictionary_path, encoding="utf-8") as f:
-            data = json.load(f)
-            # Ensure similarity_words exists
-            if "similarity_words" not in data:
-                data["similarity_words"] = []
-            return data
-    except (json.JSONDecodeError, OSError):
-        return {"similarity_words": [], "replacements": {}, "case_sensitive": {}, "patterns": []}
-
-
-def save_dictionary(data: dict[str, Any]) -> None:
-    """Save dictionary to file."""
-    dictionary_path = Config.CONFIG_DIR / "dictionary.json"
-    dictionary_path.parent.mkdir(parents=True, exist_ok=True)
-
-    with open(dictionary_path, "w", encoding="utf-8") as f:
-        json.dump(data, f, indent=2, ensure_ascii=False)
-
-
-def add_similarity_word(word: str) -> None:
-    """Add a word to the similarity dictionary."""
-    data = get_dictionary()
-    if "similarity_words" not in data:
-        data["similarity_words"] = []
-    if word not in data["similarity_words"]:
-        data["similarity_words"].append(word)
-        save_dictionary(data)
-
-
-def remove_similarity_word(word: str) -> None:
-    """Remove a word from the similarity dictionary."""
-    data = get_dictionary()
-    if word in data.get("similarity_words", []):
-        data["similarity_words"].remove(word)
-        save_dictionary(data)
+# -- re-exports: config logic -------------------------------------------------
+from .web.config_logic import (  # noqa: F401
+    CONFIG_BOOL_FIELDS,
+    CONFIG_FIELD_MAP,
+    CONFIG_STRING_FIELDS,
+    _get_env_bool,
+    _get_env_string,
+    _hotkey_status,
+    _llm_status,
+    _mask_api_key,
+    _setup_state,
+    _stt_status,
+    _text_output_status,
+    add_similarity_word,
+    build_setup_status,
+    get_current_config,
+    get_dictionary,
+    remove_similarity_word,
+    save_config,
+    save_dictionary,
+)
+
+# -- re-exports: env I/O -----------------------------------------------------
+from .web.env_io import (  # noqa: F401
+    _find_env_file,
+    get_env_path,
+    read_env_file,
+    write_env_file,
+)
+
+# -- re-exports: templates ----------------------------------------------------
+from .web.templates import (  # noqa: F401
+    ADVANCED_HTML_TEMPLATE,
+    LOGO_BASE64,
+    SETUP_HTML_TEMPLATE,
+    _load_html_template,
+    _load_logo_base64,
+)
 
 
 def create_app():
