+++ src/dicton/interfaces/web/env_io.py
@@ -1,0 +1,63 @@
+"""Environment file I/O for the Dicton configuration."""
+
+from __future__ import annotations
+
+from pathlib import Path
+
+from ...shared.config import Config
+
+
+def get_env_path() -> Path:
+    """Get the .env file path for writing (always user config dir)."""
+    return Config.CONFIG_DIR / ".env"
+
+
+def _find_env_file() -> Path | None:
+    """Find .env file - prioritize user config over system config."""
+    locations = [
+        Config.CONFIG_DIR / ".env",  # User config dir (~/.config/dicton/) - FIRST!
+        Path.cwd() / ".env",  # Current working directory
+        Path("/opt/dicton/.env"),  # System install (read-only fallback)
+    ]
+    for env_path in locations:
+        if env_path.exists():
+            return env_path
+    return None
+
+
+def read_env_file() -> dict[str, str]:
+    """Read the .env file and return as dict (checks multiple locations)."""
+    env_path = _find_env_file()
+    if env_path is None:
+        return {}
+
+    env_vars: dict[str, str] = {}
+    with open(env_path, encoding="utf-8") as f:
+        for line in f:
+            line = line.strip()
+            if not line or line.startswith("#"):
+                continue
+            if "=" in line:
+                key, _, value = line.partition("=")
+                # Strip quotes from value
+                value = value.strip().strip("\"'")
+                env_vars[key.strip()] = value
+
+    return env_vars
+
+
+def write_env_file(env_vars: dict[str, str]) -> None:
+    """Write env vars to .env file."""
+    env_path = get_env_path()
+    env_path.parent.mkdir(parents=True, exist_ok=True)
+
+    lines = ["# Dicton configuration", "# Generated by dicton --config", ""]
+
+    for key, value in sorted(env_vars.items()):
+        # Quote values with spaces
+        if " " in value or not value:
+            value = f'"{value}"'
+        lines.append(f"{key}={value}")
+
+    with open(env_path, "w", encoding="utf-8") as f:
+        f.write("\n".join(lines) + "\n")
